Fixed success + failure callbacks
