+++ app/models/spree/payu_detail.rb
@@ -1,5 +1,5 @@
 module Spree
-  class PayuDetail < ApplicationRecord
+  class PayuDetail < Spree::Base
     belongs_to :payment, class_name: 'Spree::Payment', optional: true
     belongs_to :order, class_name: 'Spree::Order'
   end
