+++ config/routes.rb
@@ -1,5 +1,5 @@
 Spree::Core::Engine.routes.draw do
   # Add your extension routes here
-  post 'spree/payu_handler/success', to: 'spree/payu_handler#success'
-  post 'spree/payu_handler/fail'
+  post 'spree/payu_handler/success', to: 'payu_handler#success'
+  post 'spree/payu_handler/fail', to: 'payu_handler#fail'
 end
